app/requirements.txt
onnxruntime>=1.14.1
chromadb==0.4.24
fastapi==0.109.2
langchain==0.1.14
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
pydantic==2.6.1
pypdf==4.0.2
python-decouple==3.8
python-docx==1.1.0
python-multipart==0.0.9
starlette==0.36.3
uvicorn==0.27.1